+++ src/addons/connect.js
@@ -8,7 +8,7 @@ import assign from 'object-assign';
 export default React => {
   const { instanceMethods, staticProperties } = createReactComponentMethods(React);
 
-  return (BaseComponent, { stores, stateGetter }) => {
+  return (BaseComponent, { stores, stateGetter, actions, actionGetter }) => {
     const ConnectedComponent = class extends React.Component {
       constructor(props, context) {
         super(props, context);
@@ -16,12 +16,19 @@ export default React => {
         this.initialize();
 
         this.state = {
-          storeState: this.connectToStores(stores, props.stateGetter)
+          storeState: this.connectToStores(stores, stateGetter),
+          actions: this.collectActions(actions, actionGetter, props)
         };
       }
 
       render() {
-        return <BaseComponent {...this.state.storeState} {...this.props} />;
+        return (
+          <BaseComponent
+            {...this.props}
+            {...this.state.storeState}
+            {...this.state.actions}
+          />
+        );
       }
     };